can run pygame from anki
will terminate when leaving reviewer

only problem so far is that when transitioning from the overview state to the review state, the mainwindow will freeze and not show the question card to the user. other than that, i am somehow able to run pygame simultaneously while moving from card to card in anki. everything renders correctly, including audio, and cards can be answered. the only downside is that anki performs noticeably worse while pygame is active. i acknowledge the fact that my testing was not very thorough, and potential problems, such as crashing, may occur in the future.